+++ pluginutils/loader/init.go
@@ -14,7 +14,7 @@ import (
 	"github.com/hashicorp/nomad/nomad/structs/config"
 	"github.com/hashicorp/nomad/plugins/base"
 	"github.com/hashicorp/nomad/plugins/shared/hclspec"
-	"github.com/hashicorp/nomad/plugins/shared/hclutils"
+	"github.com/hashicorp/nomad/pluginutils/hclutils"
 	"github.com/zclconf/go-cty/cty/msgpack"
 )
